Configure react-hot-loader in development
