+++ .babelrc
@@ -16,7 +16,10 @@
         "react",
         "stage-0"
       ],
-      "plugins": ["transform-runtime"]
+      "plugins": [
+        "transform-runtime",
+        "react-hot-loader/babel"
+      ]
     },
     "test": {
       "presets": [
