+++ package.json
@@ -50,6 +50,7 @@
     "react-dom": "^16.0.0",
     "react-frame-aware-selection-plugin": "^1.0.0",
     "react-frame-component": "^1.1.1",
+    "react-hot-loader": "^3.1.3",
     "react-portal": "^3.1.0",
     "react-router": "^2.5.1",
     "react-router-dom": "^4.1.1",
